requirements.txt
Django==2.2
django-crispy-forms==1.9.2
django-etc==1.2.0
django-hitcount==1.3.3
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
